+++ NDui/Modules/Misc/ExpRep.lua
@@ -33,8 +33,13 @@ function M:ExpBar_Update()
 		local factionID = factionData.factionID
 		if factionID and C_Reputation.IsMajorFaction(factionID) then
 			local majorFactionData = C_MajorFactions.GetMajorFactionData(factionID)
-			value = majorFactionData.renownReputationEarned or 0
-			barMin, barMax = 0, majorFactionData.renownLevelThreshold
+			local isMaxRenown = C_MajorFactions.HasMaximumRenown(factionID)
+			if isMaxRenown then
+				barMin, barMax, value = 0, 1, 1
+			else
+				value = majorFactionData.renownReputationEarned or 0
+				barMin, barMax = 0, majorFactionData.renownLevelThreshold
+			end
 		else
 			local repInfo = C_GossipInfo.GetFriendshipReputation(factionID)
 			local friendID, friendRep, friendThreshold, nextFriendThreshold = repInfo.friendshipFactionID, repInfo.standing, repInfo.reactionThreshold, repInfo.nextThreshold
@@ -118,9 +123,15 @@ function M:ExpBar_UpdateTooltip()
 		if factionID and C_Reputation.IsMajorFaction(factionID) then
 			local majorFactionData = C_MajorFactions.GetMajorFactionData(factionID)
 			name = majorFactionData.name
-			value = majorFactionData.renownReputationEarned or 0
-			barMin, barMax = 0, majorFactionData.renownLevelThreshold
 			standingtext = RENOWN_LEVEL_LABEL..majorFactionData.renownLevel
+
+			local isMaxRenown = C_MajorFactions.HasMaximumRenown(factionID)
+			if isMaxRenown then
+				barMin, barMax, value = 0, 1, 1
+			else
+				value = majorFactionData.renownReputationEarned or 0
+				barMin, barMax = 0, majorFactionData.renownLevelThreshold
+			end
 		else
 			local repInfo = C_GossipInfo.GetFriendshipReputation(factionID)
 			local friendID, friendRep, friendThreshold, nextFriendThreshold = repInfo.friendshipFactionID, repInfo.standing, repInfo.reactionThreshold, repInfo.nextThreshold