+++ HandyNotes_WarWithin/CHANGELOG.md
@@ -1,10 +1,15 @@
 # Changelog
 
+## Changed in v14
+
+* Azj-Kahet: found Silk-Spun Supplies, Concealed Contraband
+* Delves: Kriegval's Rest, Fungal Folly
+
 ## Changed in v13
 
 * Isle of Dorn: consolidated the Violet Hold Prisoner nodes for better display
 * Hallowfall: put the unknown-location rares that're tied to Defender of the Flame onto the map nearby as placeholders
-* Azj-Kahet: found the Niffen Stash, loot from Webspeaker Grik'ik, some very incomplete data on the Smelling History and Bookwork achievements
+* Azj-Kahet: found the Niffen Stash, loot from Webspeaker Grik'ik, some very incomplete data on the Smelling History and Bookworm achievements
 
 ## Changed in v12